Move AuthHandler to AuthHandler.cs
